feat: injection and pi

+++ PRIO/src/Modules/Balance/Injection/Infra/EF/Models/InjectionWaterWell.cs
@@ -6,7 +6,7 @@ namespace PRIO.src.Modules.Balance.Injection.Infra.EF.Models
     public class InjectionWaterWell
     {
         public Guid Id { get; set; }
-        public double AssignedValue { get; set; }
+        public decimal AssignedValue { get; set; }
         public User CreatedBy { get; set; }
         public DateTime MeasurementAt { get; set; }
         public WellsValues WellValues { get; set; }
